pyproject.toml
[tool.poetry]
name = "spatial-comp-1"
version = "0.1.0"
description = ""
authors = ["Ben Eidson <b.e.eidson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
